Authorization

+++ backend/src/main/java/com/example/book_n_go/controller/HallController.java
@@ -24,6 +24,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.example.book_n_go.dto.HallRequest;
 import com.example.book_n_go.dto.HallsFilterRequest;
+import com.example.book_n_go.enums.Permission;
 import com.example.book_n_go.model.Aminity;
 import com.example.book_n_go.model.Hall;
 import com.example.book_n_go.model.Workspace;
@@ -73,6 +74,9 @@ public class HallController {
     @PostMapping("/halls")
     public ResponseEntity<Hall> createHall(@RequestBody HallRequest hallRequest, @PathVariable("workspaceId") long workspaceId) {
         Workspace workspace = workspaceRepo.findById(workspaceId).get();
+        if (!workspace.getProvider().getId().equals(AuthService.getRequestUser().getId()) || !AuthService.userHasPermission(Permission.PROVIDER_WRITE)) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         if(workspace == null) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
@@ -86,6 +90,9 @@ public class HallController {
     @PutMapping("/halls/{id}")
     public ResponseEntity<Hall> updateHall(@PathVariable("id") long id, @RequestBody HallRequest hall) {
         Optional<Hall> hallData = hallRepo.findById(id);
+        if (!AuthService.userHasPermission(Permission.PROVIDER_UPDATE) || !AuthService.getRequestUser().getId().equals(hallData.get().getWorkspace().getProvider().getId())) {
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         if (hallData.isPresent()) {
             Hall _hall = hallData.get();
             _hall.setCapacity(hall.getCapacity());
@@ -105,6 +112,9 @@ public class HallController {
     @DeleteMapping("/halls/{id}")
     public ResponseEntity<HttpStatus> deleteHall(@PathVariable("id") long id) {
         try {
+            if(!AuthService.userHasPermission(Permission.PROVIDER_DELETE) || !AuthService.getRequestUser().getId().equals(hallRepo.findById(id).get().getWorkspace().getProvider().getId())) {
+              return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+            }
             hallRepo.deleteById(id);
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
         } catch (Exception e) {
