+++ backend/src/main/java/com/example/book_n_go/service/BookingService.java
@@ -14,6 +14,7 @@ import com.example.book_n_go.dto.BookingCreateRequest;
 import com.example.book_n_go.dto.BookingUpdateDurationRequest;
 import com.example.book_n_go.dto.BookingUpdateStatusRequest;
 import com.example.book_n_go.enums.Day;
+import com.example.book_n_go.enums.Permission;
 import com.example.book_n_go.enums.Status;
 import com.example.book_n_go.model.Booking;
 import com.example.book_n_go.model.HallSchedule;
@@ -132,6 +133,11 @@ public class BookingService {
             throw new IllegalArgumentException("Cannot delete booking within 48 hours of the start time");
         }
 
+        if(!((authService.userHasPermission(Permission.PROVIDER_DELETE) && authService.getRequestUser().getId().equals(bookingRepo.findById(bookingId).get().getHall().getWorkspace().getProvider().getId()))
+              || (authService.userHasPermission(Permission.CLIENT_DELETE) && authService.getRequestUser().getId().equals(bookingRepo.findById(bookingId).get().getUser().getId())))) {
+            throw new IllegalArgumentException("Unauthorized");
+        }
+
         bookingRepo.deleteById(bookingId);
     }
 
@@ -203,6 +209,13 @@ public class BookingService {
         }
         
         Booking booking = bookingRepo.findById(bookingId).get();
+
+        // Authorization
+        if(!((authService.userHasPermission(Permission.PROVIDER_UPDATE) && authService.getRequestUser().getId().equals(booking.getHall().getWorkspace().getProvider().getId()))
+            || (authService.userHasPermission(Permission.CLIENT_UPDATE) && authService.getRequestUser().getId().equals(booking.getUser().getId())))) {
+            throw new IllegalArgumentException("Unauthorized");
+        }
+
         booking.setStartTime(startTime);
         booking.setEndTime(endTime);
         double totalCost = booking.getHall().getPricePerHour() * Duration.between(booking.getStartTime(), booking.getEndTime()).toHours();
@@ -229,6 +242,12 @@ public class BookingService {
 
         if (booking.getStatus() == Status.REJECTED || booking.getStatus() == Status.CONFIRMED || booking.getStatus() == Status.CANCELED) {
             throw new IllegalArgumentException("Booking is already " + booking.getStatus());
+        }
+
+        // Authorization
+        if(!((authService.userHasPermission(Permission.PROVIDER_UPDATE) && authService.getRequestUser().getId().equals(booking.getHall().getWorkspace().getProvider().getId()))
+            || (authService.userHasPermission(Permission.CLIENT_UPDATE) && authService.getRequestUser().getId().equals(booking.getUser().getId())))) {
+            throw new IllegalArgumentException("Unauthorized");
         }
 
         booking.setStatus(status);
